Small improvements in tntpub::Client.
Use cxxtools::IOStream in client. Using it makes unblocking mode a little
easier since it marks the underlying device ready for writing when sending a
message.

Sending methods return a reference to the client so that methods can be
chained.

Add flush method to client, which forces the buffer to be written in blocking
mode.