Fix last fix
Blocking read did not work at all since the data message war reinitialized
before use, so the message was always empty.
Not reinitialization is moved to the start of the blocking read so that the
last message is kept until the user signals, that he wants to read the  next
one.